House of Wear Store Data Analysis Project
"House of Wears," a clothing store with a diverse range of products, is looking to enhance its business strategies for the year 2022. To achieve this, they aim to create a comprehensive sales report for the year and use Excel visualization tools to gain valuable insights into their customers' behavior. The objective is to understand their customer base better and identify opportunities for growth and improvement.